+++ eslint.config.ts
@@ -69,6 +69,7 @@ export default [
   {
     ignores: [
       'node_modules',
+      'plans',
       'cache',
       'coverage',
       'typechain-types',