+++ src/definitions/mantle.ts
@@ -9,4 +9,10 @@ export const mantle: Chain = {
       http: ['https://mantle-mainnet.g.alchemy.com/v2'],
     },
   },
+  contracts: {
+    ...vmantle.contracts,
+    hyperlaneMailbox: {
+      address: '0x398633D19f4371e1DB5a8EFE90468eB70B1176AA',
+    },
+  },
 }