+++ src/definitions/optimism.ts
@@ -9,6 +9,12 @@ export const optimism: Chain = {
       http: ['https://opt-mainnet.g.alchemy.com/v2'],
     },
   },
+  contracts: {
+    ...vop.contracts,
+    hyperlaneMailbox: {
+      address: '0xd4C1905BB1D26BC93DAC913e13CaCC278CdCC80D',
+    },
+  },
 }
 
 export const optimismSepolia: Chain = {
@@ -17,6 +23,12 @@ export const optimismSepolia: Chain = {
     ...vops.rpcUrls,
     alchemy: {
       http: ['https://opt-sepolia.g.alchemy.com/v2'],
+    },
+  },
+  contracts: {
+    ...vops.contracts,
+    hyperlaneMailbox: {
+      address: '0x6966b0E55883d49BFB24539356a2f8A673E02039',
     },
   },
 }